backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[argon2]
pydantic-settings
python-dotenv
aiofiles
slowapi
